+++ packages/toprf-secure-backup/src/toprfSecureBackup.test.ts
@@ -139,8 +139,7 @@ describe('toprf secret backup', function () {
           userId,
         }),
       ).rejects.toThrow('Failed to get node details');
-
-      expect(fndSpy).toHaveBeenCalled();
+      expect(fndSpy).toHaveBeenCalledTimes(1);
     });
   });
 
@@ -352,7 +351,7 @@ describe('toprf secret backup', function () {
           /Node details arrays .* must have equal lengths/iu,
         );
 
-        expect(fndSpy).toHaveBeenCalledTimes(1);
+        expect(fndSpy).toHaveBeenCalledTimes(2);
       });
     });
   });
@@ -887,11 +886,7 @@ describe('toprf secret backup', function () {
             curAuthKeyPair: newEncKeyResult2.authKeyPair,
             maxPwChainLength: 1,
           }),
-        ).rejects.toThrow(
-          TOPRFError.couldNotFetchPassword(
-            'Exceeded maximum password chain length',
-          ),
-        );
+        ).rejects.toThrow(TOPRFError.maxKeyChainLengthExceeded());
       });
       it(`should be able to change encryption key and recover password for ${fmtGroupedConnId(groupedConnId) ? 'aggregate' : 'normal'} verifier using pregenerated OPRF key`, async function () {
         const secretData = utf8ToBytes('test-secret-data-for-key-change');
@@ -1065,11 +1060,7 @@ describe('toprf secret backup', function () {
             curAuthKeyPair: newEncKeyResult2.authKeyPair,
             maxPwChainLength: 1,
           }),
-        ).rejects.toThrow(
-          TOPRFError.couldNotFetchPassword(
-            'Exceeded maximum password chain length',
-          ),
-        );
+        ).rejects.toThrow(TOPRFError.maxKeyChainLengthExceeded());
       });
     });