+++ packages/toprf-secure-backup/src/toprfSecureBackup.ts
@@ -100,6 +100,11 @@ export class ToprfSecureBackup implements IToprfSecureBackup {
     this.#nodeDetailsOverride = params.nodeDetailsOverride;
     this.#keyDeriver = params.keyDeriver;
     this.#fetchMetadataAccessCreds = params.fetchMetadataAccessCreds;
+    // pre-fetch node details to speed up the first call to authenticate
+    // fnd package internally caches results so preloading will speed up subsequent calls.
+    this.#getNodeDetails().catch((error) => {
+      console.error('Error fetching node details', error);
+    });
   }
 
   /**
@@ -673,9 +678,7 @@ export class ToprfSecureBackup implements IToprfSecureBackup {
       }
     }
 
-    throw TOPRFError.couldNotFetchPassword(
-      'Exceeded maximum password chain length',
-    );
+    throw TOPRFError.maxKeyChainLengthExceeded();
   }
 
   /**